requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
plotly
shap
jupyter
ipykernel
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1